refactor-#156: move customer service to services
