+++ src/microservices/customer/graphql/customer-mutation.resolver.ts
@@ -1,10 +1,11 @@
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
+import { UseGuards } from '@nestjs/common';
+
+import { GqlJwtAuthGuard } from '@auth/graphql/guards';
 import { GqlCustomer } from './models/customer-gql.model';
-import { CustomerService } from '../customer.service';
+import { CustomerService } from '../services';
 import { CreateCustomerProfileInput } from './input/create-customer-profile.input';
 import { UpdateCustomerProfileInput } from './input/update-customer-profile.input';
-import { UseGuards } from '@nestjs/common';
-import { GqlJwtAuthGuard } from '@auth/graphql/guards';
 
 @Resolver((of) => GqlCustomer)
 export class CustomerMutationResolver {
