+++ src/microservices/customer/graphql/customer-query.resolver.ts
@@ -1,8 +1,9 @@
 import { NotFoundException, UseGuards } from '@nestjs/common';
 import { Args, Resolver, Int, Query } from '@nestjs/graphql';
-import { GqlCustomer } from './models/customer-gql.model';
-import { CustomerService } from '../customer.service';
+
 import { GqlJwtAuthGuard } from '@auth/graphql/guards';
+import { GqlCustomer } from './models/customer-gql.model';
+import { CustomerService } from '../services';
 
 @Resolver((of) => GqlCustomer)
 export class CustomerQueryResolver {
