+++ src/microservices/customer/rest/customer.controller.ts
@@ -8,8 +8,9 @@ import {
   Delete,
   Put,
 } from '@nestjs/common';
-import { CustomerService } from '../customer.service';
+
 import { RestJwtAuthGuard } from '@auth/rest/guards';
+import { CustomerService } from '../services';
 import { CreateCustomerProfileDto } from './dto/create-customer-profile.dto';
 import { RestCustomer } from './models/customer-rest.model';
 import { UpdateCustomerProfileDto } from './dto/update-customer-profile.dto';
